# backend/docs/swagger.yaml
basePath: /api
definitions:
  api.errorResponse:
    properties:
      message:
        type: string
    type: object
  internal.CreateCommentRequest:
    properties:
      id:
        type: integer
      text:
        type: string
    required:
    - id
    - text
    type: object
  internal.CreateCommentResponse:
    properties:
      id:
        type: integer
    type: object
  internal.CreateFollowRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal.CreateLikeRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  internal.CreateUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      fullname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - fullname
    - password
    - username
    type: object
  internal.CreateUserResponse:
    properties:
      id:
        type: integer
      token:
        type: string
    type: object
  internal.DeleteCommentRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal.DeleteFollowRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal.DeleteLikeRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  internal.GetProjectResponse:
    properties:
      comments_count:
        type: integer
      created_at:
        type: string
      description:
        type: string
      folder:
        type: string
      id:
        type: integer
      is_liked:
        type: boolean
      likes_count:
        type: integer
      name:
        type: string
      readme:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal.GetUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      followers_count:
        type: integer
      fullname:
        type: string
      id:
        type: integer
      is_followed:
        type: boolean
      projects:
        items:
          $ref: '#/definitions/internal.GetProjectResponse'
        type: array
      username:
        type: string
    type: object
...
  internal.UpdateUserRequest:
    properties:
      bio:
        type: string
      fullname:
        type: string
    type: object
info:
  contact: {}
  title: WebPortfolio API
  version: "1.0"
paths:
...
  /comment:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.DeleteCommentRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: comment not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Delete comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: comments
          schema:
            items:
              properties:
                created_at:
                  type: string
                id:
                  type: integer
                text:
                  type: string
                user:
                  properties:
                    fullname:
                      type: string
                    id:
                      type: integer
                    username:
                      type: string
                  type: object
              type: object
            type: array
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: project not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: comment
          schema:
            $ref: '#/definitions/internal.CreateCommentResponse'
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: project not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Create comment
      tags:
      - comment
  /feed:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: feed
          schema:
            items:
              $ref: '#/definitions/internal.GetProjectResponse'
            type: array
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Get feed
      tags:
      - feed
  /follow:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.DeleteFollowRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: user is not followed
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Delete follow
      tags:
      - follow
    post:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateFollowRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: user is already followed
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Create follow
      tags:
      - follow
  /followers:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: follows
          schema:
            items:
              properties:
                fullname:
                  type: string
                id:
                  type: integer
                username:
                  type: string
              type: object
            type: array
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get followers
      tags:
      - follow
  /following:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: follows
          schema:
            items:
              properties:
                fullname:
                  type: string
                id:
                  type: integer
                username:
                  type: string
              type: object
            type: array
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get following
      tags:
      - follow
  /like:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.DeleteLikeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: project not found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: project is not liked
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Delete like
      tags:
      - like
    post:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateLikeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: project not found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: project is already liked
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Create like
      tags:
      - like
...
  /trending:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: trending
          schema:
            items:
              $ref: '#/definitions/internal.GetProjectResponse'
            type: array
      security:
      - Bearer: []
      summary: Get trending
      tags:
      - feed
...
  /user:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: projects
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/internal.GetUserResponse'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Get user
      tags:
      - user
    patch:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.UpdateUserRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: not authorized
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateUserRequest'
      responses:
        "201":
          description: user
          schema:
            $ref: '#/definitions/internal.CreateUserResponse'
        "400":
          description: validation failed
          schema:
            $ref: '#/definitions/api.errorResponse'
        "409":
          description: email already exists
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Create user
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"